// src/components/types.ts
export type InstanceId = `${number}-${number}`
export type Label = "Do" | "Be" | "Feel" | "Concern" | "Who";
// ...
export type TreeGoal = {
    id: number;
    content: string;
    type: Label;
    instanceId: InstanceId;
    children?: TreeGoal[];
    color?: string;
};